requirements.txt
SQLAlchemy>=2.0.0
fastapi>=0.88.0
uvicorn>=0.20.0
psycopg2-binary==2.8.6
passlib==1.7.4
PyJWT
databases==0.4.1
asyncpg==0.21.0
alembic>=1.8.0
bcrypt==4.0.1
fastapi_jwt_auth
jinja2
pathlib
pydantic>=1.10.0
pydantic-settings
#python-dotenv
python-multipart
python-jose[cryptography]
uuid
cloud-sql-python-connector==1.2.4
functions-framework==3.3.0
pg8000==1.29.8
fastapi_limiter
celery>=5.3.0
redis>=4.5.0
pika==1.3.2
tweepy==4.14.0
python-multipart
sqlparse
psycopg2
flower
